Restrict Value utilities to crate
